feat: add ContextMenuAction protocol for generic useage
