+++ TesserCube/Helper/SharePublicKeyActivity.swift
